Added FNF modified version for Supergames
Added FNF modified version for Supergames, in accordance to the license of Friday Night Funkin':

https://github.com/ninjamuffin99/Funkin

Of course, code respects the original license of Friday Night Funkin', as established by its creators.

Had to change the screen that appears before the main menu, that said that the version isn't the latest one, and also had to remove the Debug information that appeared when playing Week 6.